+++ CHANGELOG.md
@@ -5,13 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
 
-### Added
+## Added
 
 - Add readme and changelog
 
-### Changed
+## Changed
 
 - Corrected the letters for `exit`.
 - Use descriptive glyph names for easy understanding
+- Fixed styles
+- Renamed slabsprak.js to script.js
+- Refactored script.js